refactor: move all tests into qpc/tests directory
