+++ qpc/tests/cred/test_cred_show.py
@@ -10,7 +10,7 @@ import requests_mock
 
 from qpc.cred import CREDENTIAL_URI
 from qpc.cred.show import CredShowCommand
-from qpc.tests_utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
+from qpc.tests.utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
 from qpc.utils import get_server_location, write_server_config
