+++ qpc/tests/insights/test_insights_auth.py
