+++ qpc/tests/insights/test_insights_configure.py
