+++ qpc/tests/scan/test_scan_add.py
@@ -15,7 +15,7 @@ from qpc.request import CONNECTION_ERROR_MSG
 from qpc.scan import SCAN_URI
 from qpc.scan.add import ScanAddCommand
 from qpc.source import SOURCE_URI
-from qpc.tests_utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
+from qpc.tests.utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
 from qpc.utils import get_server_location, write_server_config
