+++ qpc/tests/scan/test_scan_cancel.py
@@ -12,19 +12,19 @@ import requests_mock
 from qpc import messages
 from qpc.request import CONNECTION_ERROR_MSG
 from qpc.scan import SCAN_JOB_URI
-from qpc.scan.restart import ScanRestartCommand
-from qpc.tests_utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
+from qpc.scan.cancel import ScanCancelCommand
+from qpc.tests.utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
 from qpc.utils import get_server_location, write_server_config
 
 
-class TestScanRestartCli:
-    """Class for testing the scan restart commands for qpc."""
+class TestScanCancelCli:
+    """Class for testing the scan cancel commands for qpc."""
 
     def _init_command(self):
         """Initialize command."""
         argument_parser = ArgumentParser()
         subparser = argument_parser.add_subparsers(dest="subcommand")
-        return ScanRestartCommand(subparser)
+        return ScanCancelCommand(subparser)
 
     def setup_method(self, _test_method):
         """Create test setup."""
@@ -33,6 +33,7 @@ class TestScanRestartCli:
         # SourceEditCommand instance modifies req_path on the fly. This seems to be a
         # code smell to me, but I'm choosing to ignore it for now
         self.command = self._init_command()
+
         write_server_config(DEFAULT_CONFIG)
         # Temporarily disable stderr for these tests, CLI errors clutter up
         # nosetests command.
@@ -44,10 +45,10 @@ class TestScanRestartCli:
         # Restore stderr
         sys.stderr = self.orig_stderr
 
-    def test_restart_scan_ssl_err(self):
-        """Testing the restart scan command with a connection error."""
+    def test_cancel_scan_ssl_err(self):
+        """Testing the cancel scan command with a connection error."""
         scan_out = StringIO()
-        url = get_server_location() + SCAN_JOB_URI + "1/restart/"
+        url = get_server_location() + SCAN_JOB_URI + "1/cancel/"
         with requests_mock.Mocker() as mocker:
             mocker.put(url, exc=requests.exceptions.SSLError)
 
@@ -57,10 +58,10 @@ class TestScanRestartCli:
                     self.command.main(args)
                     assert scan_out.getvalue() == CONNECTION_ERROR_MSG
 
-    def test_restart_scan_conn_err(self):
-        """Testing the restart scan command with a connection error."""
+    def test_cancel_scan_conn_err(self):
+        """Testing the cancel scan command with a connection error."""
         scan_out = StringIO()
-        url = get_server_location() + SCAN_JOB_URI + "1/restart/"
+        url = get_server_location() + SCAN_JOB_URI + "1/cancel/"
         with requests_mock.Mocker() as mocker:
             mocker.put(url, exc=requests.exceptions.ConnectTimeout)
 
@@ -70,10 +71,10 @@ class TestScanRestartCli:
                     self.command.main(args)
                     assert scan_out.getvalue() == CONNECTION_ERROR_MSG
 
-    def test_restart_scan_internal_err(self):
-        """Testing the restart scan command with an internal error."""
+    def test_cancel_scan_internal_err(self):
+        """Testing the cancel scan command with an internal error."""
         scan_out = StringIO()
-        url = get_server_location() + SCAN_JOB_URI + "1/restart/"
+        url = get_server_location() + SCAN_JOB_URI + "1/cancel/"
         with requests_mock.Mocker() as mocker:
             mocker.put(url, status_code=500, json={"error": ["Server Error"]})
 
@@ -83,14 +84,14 @@ class TestScanRestartCli:
                     self.command.main(args)
                     assert scan_out.getvalue() == "Server Error"
 
-    def test_restart_scan_data(self, caplog):
-        """Testing the restart scan command successfully with stubbed data."""
-        url = get_server_location() + SCAN_JOB_URI + "1/restart/"
+    def test_cancel_scan_data(self, caplog):
+        """Testing the cancel scan command successfully with stubbed data."""
+        url = get_server_location() + SCAN_JOB_URI + "1/cancel/"
         with requests_mock.Mocker() as mocker:
             mocker.put(url, status_code=200, json=None)
 
             args = Namespace(id="1")
             with caplog.at_level(logging.INFO):
                 self.command.main(args)
-                expected_message = messages.SCAN_RESTARTED % "1"
+                expected_message = messages.SCAN_CANCELED % "1"
                 assert expected_message in caplog.text
