+++ qpc/tests/scan/test_scan_job.py
@@ -12,7 +12,7 @@ from qpc import messages
 from qpc.request import CONNECTION_ERROR_MSG
 from qpc.scan import SCAN_JOB_URI, SCAN_URI
 from qpc.scan.job import ScanJobCommand
-from qpc.tests_utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
+from qpc.tests.utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
 from qpc.utils import get_server_location, write_server_config
