+++ qpc/tests/scan/test_scan_start.py
@@ -11,7 +11,7 @@ from qpc import messages
 from qpc.cli import CLI
 from qpc.scan import SCAN_URI
 from qpc.scan.start import ScanStartCommand
-from qpc.tests_utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
+from qpc.tests.utilities import DEFAULT_CONFIG, HushUpStderr, redirect_stdout
 from qpc.utils import get_server_location, write_server_config
